% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_nichediv_dependencies.R
\name{install_nichediv_dependencies}
\alias{install_nichediv_dependencies}
\title{Install external dependencies for NicheDiv}
\usage{
install_nichediv_dependencies(
  install_climatena = TRUE,
  install_whitebox = TRUE,
  install_data_table = TRUE,
  climatena_url = "https://zenodo.org/records/17401570/files/ClimateNAr.zip?download=1",
  timeout = 800,
  force = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{install_climatena}{Logical; whether to install ClimateNAr.}

\item{install_whitebox}{Logical; whether to install whitebox.}

\item{install_data_table}{Logical; whether to install data.table.}

\item{climatena_url}{Download URL for the ClimateNAr Windows binary zip.}

\item{timeout}{Download timeout in seconds.}

\item{force}{Logical; reinstall even if package is already installed.}

\item{verbose}{Logical; print progress messages.}
}
\value{
Invisibly returns \code{TRUE} when finished.
}
\description{
Installs optional packages used by selected NicheDiv workflows,
including packages not available on CRAN such as ClimateNAr.
}
